dbus: hostname: move out of networkmanager context
The hostname DBus interface is actually provided by systemd-hostnamed and
not by NetworkManager.
Having it in the networkmanager context is thus technically not correct.
Move it into its own context.
