+++ src/dbus/networkmanager/mod.rs
@@ -24,7 +24,6 @@ use crate::broker::{BrokerBuilder, Topic};
 use crate::led::BlinkPattern;
 
 mod devices;
-mod hostname;
 
 // All of the following includes are not used in demo_mode.
 // Put them inside a mod so we do not have to decorate each one with
@@ -237,7 +236,6 @@ impl<'a> IpStream<'a> {
 }
 
 pub struct Network {
-    pub hostname: Arc<Topic<String>>,
     pub bridge_interface: Arc<Topic<Vec<String>>>,
     pub dut_interface: Arc<Topic<LinkInfo>>,
     pub uplink_interface: Arc<Topic<LinkInfo>>,
@@ -246,7 +244,6 @@ pub struct Network {
 impl Network {
     fn setup_topics(bb: &mut BrokerBuilder) -> Self {
         Self {
-            hostname: bb.topic_ro("/v1/tac/network/hostname", None),
             bridge_interface: bb.topic_ro("/v1/tac/network/interface/tac-bridge", None),
             dut_interface: bb.topic_ro("/v1/tac/network/interface/dut", None),
             uplink_interface: bb.topic_ro("/v1/tac/network/interface/uplink", None),
@@ -262,7 +259,6 @@ impl Network {
     ) -> Self {
         let this = Self::setup_topics(bb);
 
-        this.hostname.set("lxatac".to_string());
         this.bridge_interface.set(vec![String::from("192.168.1.1")]);
         this.dut_interface.set(LinkInfo {
             speed: 0,
@@ -284,26 +280,6 @@ impl Network {
         led_uplink: Arc<Topic<BlinkPattern>>,
     ) -> Self {
         let this = Self::setup_topics(bb);
-
-        {
-            let conn = conn.clone();
-            let hostname_topic = this.hostname.clone();
-            async_std::task::spawn(async move {
-                let proxy = hostname::HostnameProxy::new(&conn).await.unwrap();
-
-                let mut stream = proxy.receive_hostname_changed().await;
-
-                if let Ok(h) = proxy.hostname().await {
-                    hostname_topic.set(h);
-                }
-
-                while let Some(v) = stream.next().await {
-                    if let Ok(h) = v.get().await {
-                        hostname_topic.set(h);
-                    }
-                }
-            });
-        }
 
         {
             let conn = conn.clone();
